sprig-config-module/pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sprig-config"
version = "0.2.0-RC1"
description = "Spring-like deep merge configuration loader for Python"
authors = ["Derik G. Wilson"]
readme = "README.md"
license = "MIT"
# Tell Poetry where your package lives (src layout)
packages = [{ include = "sprigconfig", from = "src" }]

[[tool.poetry.source]]
name = "gitlab-pypi"
url = "https://gitlab.com/api/v4/projects/72230105/packages/pypi"
priority = "supplemental"

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
PyYAML = ">=6.0.2"
cryptography = "^46.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.0"
pytest = "^8.4.2"
pytest-cov = ">=6.2.1"
coverage = ">=7.10.1"
colorama = ">=0.4.6"
iniconfig = ">=2.1.0"
packaging = ">=25.0"
pluggy = ">=1.6.0"
Pygments = ">=2.19.2"
pyproject-hooks = ">=1.2.0"
python-dotenv = "^1.1.1"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = "tests"
pythonpath = "src"
markers = [
  "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
  "crypto: mark test as requiring real crypto key"
]
